JavaScript: Refactor utils.crossFade(…)
Instead of detaching both elements and sliding the parent element, detach just the element to hide. We change the parent element's height to the height of the element to show by either growing the (still attached) element to show, or by shrinking a new placeholder element. The sliding transition isn't synchronized with the fading, so we can change the element's height mid-transition without worrying about the cross fading.