JavaScript: Move PicoAdmin.askFileName() to PicoAdminModule
File name modals are always module-specific, but not necesarily related to the currently active module. For example, a user might want to create a content page while still viewing the landing page. In this case landing is the current module, but the file name modal is actually related to the PicoContentAdmin module.

No functional changes